Survival Boundary Box and Cleaned up Code
(+) The survival boundary box is a square area that the agents are required to stay within to survive.
(+) ScaleController.cs script changes the size of the survival boundary box dynamically
(*) I cleaned up the code in most if not all of the scripts.. removing unnecessary lines and adding easy-to-read comments.